Add files via upload
This code maps the angular position of the upper-limb exoskeleton to the corresponding actuation gain needed for gravity compensation. The data was aquired using very slow sinusoidal inputs to replicate non existing dynamics during movements. 

We will compare this data to the GC model we ourself calculated using Euler-Lagrangian method and try to predict the real actuation gain.